LPD-37147 Move/SF
